requirements.txt
-e .
colour-runner==0.0.4
coverage==4.2
dj-database-url==0.4.1
django==1.9.9
factory-boy==2.7.0
flake8-import-order==0.9.2
flake8==3.0.4
incuna-test-utils==6.4.1
psycopg2==2.6.2
